+++ common/src/main/java/com/illusivesoulworks/customfov/mixin/core/OptionsMixin.java
@@ -1,6 +1,6 @@
 package com.illusivesoulworks.customfov.mixin.core;
 
-import com.illusivesoulworks.customfov.CustomFovMod;
+import com.illusivesoulworks.customfov.CustomFovProfiles;
 import net.minecraft.client.Options;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -12,6 +12,6 @@ public class OptionsMixin {
 
   @Inject(at = @At("TAIL"), method = "save")
   private void customfov$save(CallbackInfo ci) {
-    CustomFovMod.saveProfiles();
+    CustomFovProfiles.saveProfiles();
   }
 }